Rename Event -> Persistable and reorganize code
