+++ src/lib.rs
@@ -17,7 +17,6 @@ mod aggregate;
 mod state;
 
 pub mod bus;
-pub mod event;
 pub mod handler;
 pub mod manager;
 pub mod store;
